# cmd/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateAuthorDTO:
    properties:
      bio:
        $ref: '#/definitions/dto.TranslatedTextDTO'
      contact:
        properties:
          email:
            example: vovka@sosijopa.com
            type: string
          links:
            properties:
              behance:
                example: vovka_art
                type: string
              facebook:
                example: vovka_fb
                type: string
              instagram:
                example: vovka_insta
                type: string
              linkedin:
                example: vovka_linkedin
                type: string
              pinterest:
                example: vovka_pins
                type: string
              telegram:
                example: vovka_tele
                type: string
              twitter:
                example: vovka_tweet
                type: string
              vkontakte:
                example: vovka_vk
                type: string
              whatsapp:
                example: "+5553535"
                type: string
              youtube:
                example: vovka_channel
                type: string
            type: object
          phone:
            example: "+5553535"
            type: string
        type: object
      is_active:
        example: true
        type: boolean
      name:
        $ref: '#/definitions/dto.TranslatedTextDTO'
    required:
    - name
    type: object
...
  dto.TranslatedTextDTO:
    properties:
      en:
        example: Tung, tung, tung, tung, tung, tung, tung, tung, tung, Sahur.
        type: string
      es:
        example: Golybini Shpionini
        type: string
      ru:
        example: Тун, тун, тун, Сахур.
        type: string
    type: object
...
  dto.UpdateAuthorDTO:
    properties:
      bio:
        $ref: '#/definitions/dto.TranslatedTextDTO'
      contact:
        properties:
          email:
            example: vovka@sosijopa.com
            type: string
          links:
            properties:
              behance:
                example: vovka_art
                type: string
              facebook:
                example: vovka_fb
                type: string
              instagram:
                example: vovka_insta
                type: string
              linkedin:
                example: vovka_linkedin
                type: string
              pinterest:
                example: vovka_pins
                type: string
              telegram:
                example: vovka_tele
                type: string
              twitter:
                example: vovka_tweet
                type: string
              vkontakte:
                example: vovka_vk
                type: string
              whatsapp:
                example: "+5553535"
                type: string
              youtube:
                example: vovka_channel
                type: string
            type: object
          phone:
            example: "+5553535"
            type: string
        type: object
      is_active:
        example: false
        type: boolean
      name:
        $ref: '#/definitions/dto.TranslatedTextDTO'
    type: object
...
  entity.Author:
    properties:
      bio:
        $ref: '#/definitions/entity.TranslatedText'
      contact:
        $ref: '#/definitions/entity.ContactInfo'
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        $ref: '#/definitions/entity.TranslatedText'
      position:
        example: 1
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
  entity.ContactInfo:
    properties:
      email:
        example: vovka@sosijopa.com
        type: string
      links:
        $ref: '#/definitions/entity.SocialLink'
      phone:
        example: "+5553535"
        type: string
    type: object
...
  entity.SocialLink:
    properties:
      behance:
        example: vovka_art
        type: string
      facebook:
        example: vovka_fb
        type: string
      instagram:
        example: vovka_insta
        type: string
      linkedin:
        example: vovka_linkedin
        type: string
      pinterest:
        example: vovka_pins
        type: string
      telegram:
        example: vovka_tele
        type: string
      twitter:
        example: vovka_tweet
        type: string
      vkontakte:
        example: vovka_vk
        type: string
      whatsapp:
        example: "+5553535"
        type: string
      youtube:
        example: vovka_channel
        type: string
    type: object
  entity.TranslatedText:
    properties:
      en:
        example: Tung, tung, tung, tung, tung, tung, tung, tung, tung, Sahur.
        type: string
      es:
        example: Golybini Shpionini
        type: string
      ru:
        example: Тун, тун, тун, Сахур.
        type: string
    type: object
host: localhost:8010
info:
  contact: {}
  description: Backend для галереи anastasii_gofman
  title: Art Gallery API
  version: "1.0"
paths:
...
  /api/authors:
    get:
      consumes:
      - application/json
      description: Получаем всех авторов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/entity.Author'
              type: array
            type: array
      summary: Get all authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: С кайфом создаем автора, все поля необязательные кроме имени, ну
        и в mail пихать нужно mail нормальный, но можно ничего не пихать
      parameters:
      - description: Author data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAuthorDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Author'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create author
      tags:
      - Authors
  /api/authors/:id:
    patch:
      consumes:
      - application/json
      description: Обновляем автора частично, все поля необязательные, не передал
        = осталось прежним
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAuthorDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Author'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Partial update author
      tags:
      - Authors
  /api/authors/{id}:
    get:
      consumes:
      - application/json
      description: Получаем автора по ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Get author by ID
      tags:
      - Authors
    put:
      consumes:
      - application/json
      description: Обновляем автора полностью, все поля необязательные, но сотрется
        вся инфа, которая была раньше и которая не передана
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      - description: Author data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAuthorDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Author'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update author
      tags:
      - Authors
...
swagger: "2.0"